+++ database/migrations/2024_12_05_205453_create_genre_movie_pivots_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('genre_movie_pivots', function (Blueprint $table) {
             $table->id();
-            $table->integer('genre_id')->nullable();
-            $table->integer('movie_id')->nullable();
+            $table->unsignedBigInteger('genre_id')->nullable();
+            $table->unsignedBigInteger('movie_id')->nullable();
 
             $table->timestamps();
